+++ engine/assets/ui/image.fs
@@ -1,0 +1,11 @@
+#version 330 core
+
+out vec4 out_color;
+
+uniform sampler2D image;
+in vec2 texCoord;
+
+void main()
+{
+    out_color = texture(image, texCoord);
+}